+++ src/assets/ba_data/python/babase/_accountv2.py
@@ -13,7 +13,7 @@ from efro.error import CommunicationError
 from efro.call import CallbackSet
 from bacommon.login import LoginType
 
-from babase._logging import accountv2log
+from babase._logging import accountlog
 import _babase
 
 if TYPE_CHECKING:
@@ -103,6 +103,12 @@ class AccountV2Subsystem:
         :meta private:
         """
         assert _babase.in_logic_thread()
+
+        # Inform the base layer of new names/etc.
+        if account is not None:
+            _babase.set_account_sign_in_state(True, account.tag)
+        else:
+            _babase.set_account_sign_in_state(False)
 
         # Fire any registered callbacks.
         for call in self.on_primary_account_changed_callbacks.getcalls():
@@ -280,7 +286,7 @@ class AccountV2Subsystem:
         # generally this means the user has explicitly signed in/out or
         # switched accounts within that back-end.
         if prev_state != new_state:
-            accountv2log.debug(
+            accountlog.debug(
                 'Implicit state changed (%s -> %s);'
                 ' will update app sign-in state accordingly.',
                 prev_state,
@@ -324,7 +330,7 @@ class AccountV2Subsystem:
             if self._implicit_signed_in_adapter is None:
                 # If implicit back-end has signed out, we follow suit
                 # immediately; no need to wait for network connectivity.
-                accountv2log.debug(
+                accountlog.debug(
                     'Signing out as result of implicit state change...',
                 )
                 plus.accounts.set_primary_credentials(None)
@@ -343,7 +349,7 @@ class AccountV2Subsystem:
                 # switching accounts via the back-end). NOTE: should
                 # test case where we don't have connectivity here.
                 if plus.cloud.is_connected():
-                    accountv2log.debug(
+                    accountlog.debug(
                         'Signing in as result of implicit state change...',
                     )
                     self._implicit_signed_in_adapter.sign_in(
@@ -376,7 +382,7 @@ class AccountV2Subsystem:
             and not signed_in_v2
             and self._implicit_signed_in_adapter is not None
         ):
-            accountv2log.debug(
+            accountlog.debug(
                 'Signing in due to on-launch-auto-sign-in...',
             )
             self._can_do_auto_sign_in = False  # Only ATTEMPT once